codigo/archived/label/image_to_mask/requirements.txt
pillow
numpy
IPython